requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
pytest>=7.0
pytest-django>=4.0
gunicorn>=20.1.0